Factor out authenticator discovery to controller concern
